backend/requirements.txt
# Myk Raws Legal RAG Backend Dependencies
# Python 3.11+

# === FastAPI & Web Framework ===
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# === Database ===
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# === Authentication & Security ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# === OpenAI & Embeddings ===
openai==1.10.0
tiktoken==0.5.2

# === Retrieval & Search ===
faiss-cpu==1.9.0.post1  # Updated for Python 3.12 compatibility
rank-bm25==0.2.2
sentence-transformers==3.3.1  # Updated to latest stable

# === NLP & Text Processing ===
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
nltk==3.8.1

# === Data Processing ===
pandas==2.2.0
numpy==1.26.3

# === Web Scraping (for data ingestion) ===
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0

# === Monitoring & Logging ===
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.40.0

# === Testing ===
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# === Development Tools ===
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# === Utilities ===
python-dateutil==2.8.2
pytz==2024.1
